consolidate insert node
