+++ src/agdb/query_builder.rs
@@ -10,12 +10,6 @@ mod insert_edges_from;
 mod insert_edges_from_to;
 mod insert_edges_values;
 mod insert_node;
-mod insert_node_alias;
-mod insert_node_values;
-mod insert_nodes;
-mod insert_nodes_aliases;
-mod insert_nodes_count;
-mod insert_nodes_values;
 mod insert_values;
 mod insert_values_ids;
 mod insert_values_multi;
