--- src/agdb/query_builder/insert_node_values.rs
@@ -1,9 +1,0 @@
-use crate::query::insert_nodes_query::InsertNodesQuery;
-
-pub struct InsertNodeValues(pub InsertNodesQuery);
-
-impl InsertNodeValues {
-    pub fn query(self) -> InsertNodesQuery {
-        self.0
-    }
-}
