--- src/agdb/query_builder/insert_nodes_values.rs
@@ -1,9 +1,0 @@
-use crate::query::insert_nodes_query::InsertNodesQuery;
-
-pub struct InsertNodesValues(pub InsertNodesQuery);
-
-impl InsertNodesValues {
-    pub fn query(self) -> InsertNodesQuery {
-        self.0
-    }
-}
